refactor: 패키지 구조 변경
기존 계층형 구조에서 헥사고날 구조로 패키지 구조 변경
