+++ src/main/java/antigravity/product/AntigravityApplication.java
@@ -1,4 +1,4 @@
-package antigravity;
+package antigravity.product;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
